+++ openapi-core/pom.xml
@@ -8,7 +8,6 @@
     <version>1.0-SNAPSHOT</version>
   </parent>
 
-  <groupId>com.github.emm035</groupId>
   <artifactId>openapi-core</artifactId>
 
   <name>OpenAPI Core</name>
