+++ openapi-schema-generator/src/main/java/com/github/emm035/openapi/schema/generator/SchemaGeneratorModule.java
@@ -5,13 +5,13 @@ import com.fasterxml.jackson.databind.JavaType;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.fasterxml.jackson.databind.type.TypeFactory;
 import com.github.emm035.openapi.core.v3.schemas.Schema;
-import com.github.emm035.openapi.schema.generator.annotations.Internal;
+import com.github.emm035.openapi.schema.generator.annotations.Extension;
 import com.github.emm035.openapi.schema.generator.annotations.RefPrefix;
-import com.github.emm035.openapi.schema.generator.assisted.Extension;
-import com.github.emm035.openapi.schema.generator.assisted.SubTypeGenerator;
+import com.github.emm035.openapi.schema.generator.annotations.internal.Internal;
 import com.github.emm035.openapi.schema.generator.base.TypeUtils;
 import com.github.emm035.openapi.schema.generator.extension.PropertyExtension;
 import com.github.emm035.openapi.schema.generator.extension.SchemaExtension;
+import com.github.emm035.openapi.schema.generator.nested.SubTypeGenerator;
 import com.github.emm035.openapi.schema.generator.visitors.ArrayFormatVisitor;
 import com.github.emm035.openapi.schema.generator.visitors.BooleanFormatVisitor;
 import com.github.emm035.openapi.schema.generator.visitors.MapFormatVisitor;
