+++ openapi-schema-generator/src/main/java/com/github/emm035/openapi/schema/generator/base/Schemas.java
@@ -5,8 +5,8 @@ import com.fasterxml.jackson.databind.ObjectMapper;
 import com.github.emm035.openapi.core.v3.references.Ref;
 import com.github.emm035.openapi.core.v3.references.Referenceable;
 import com.github.emm035.openapi.core.v3.schemas.Schema;
-import com.github.emm035.openapi.schema.generator.annotations.Internal;
 import com.github.emm035.openapi.schema.generator.annotations.RefPrefix;
+import com.github.emm035.openapi.schema.generator.annotations.internal.Internal;
 import com.github.emm035.openapi.schema.generator.assisted.RefFactory;
 import com.github.emm035.openapi.schema.generator.exceptions.SchemaGenerationException;
 import com.google.common.base.Preconditions;
