+++ openapi-schema-generator/src/main/java/com/github/emm035/openapi/schema/generator/exceptions/SchemaGenerationException.java
@@ -13,15 +13,4 @@ public class SchemaGenerationException extends Exception {
   public SchemaGenerationException(Throwable cause) {
     super(cause);
   }
-
-  public SchemaGenerationException(
-    String message,
-    Throwable cause,
-    boolean enableSuppression,
-    boolean writableStackTrace
-  ) {
-    super(message, cause, enableSuppression, writableStackTrace);
-  }
-
-  public SchemaGenerationException() {}
 }
