+++ openapi-schema-generator/src/main/java/com/github/emm035/openapi/schema/generator/nested/NestedSchemaGenerator.java
@@ -1,12 +1,12 @@
-package com.github.emm035.openapi.schema.generator.assisted;
+package com.github.emm035.openapi.schema.generator.nested;
 
 import com.fasterxml.jackson.databind.JavaType;
 import com.fasterxml.jackson.databind.JsonMappingException;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.github.emm035.openapi.core.v3.references.Referenceable;
 import com.github.emm035.openapi.core.v3.schemas.Schema;
-import com.github.emm035.openapi.schema.generator.SchemaGeneratorVisitorWrapper;
-import com.github.emm035.openapi.schema.generator.annotations.Internal;
+import com.github.emm035.openapi.schema.generator.annotations.internal.Internal;
+import com.github.emm035.openapi.schema.generator.visitors.SchemaGeneratorVisitorWrapper;
 import com.google.inject.Inject;
 
 public class NestedSchemaGenerator {
