+++ openapi-schema-generator/src/main/java/com/github/emm035/openapi/schema/generator/result/SchemaResult.java
@@ -1,6 +1,8 @@
 package com.github.emm035.openapi.schema.generator.result;
 
+import com.fasterxml.jackson.annotation.JsonCreator;
 import com.fasterxml.jackson.annotation.JsonGetter;
+import com.fasterxml.jackson.annotation.JsonProperty;
 import com.github.emm035.openapi.core.v3.schemas.Schema;
 import java.util.Map;
 
@@ -13,7 +15,11 @@ public class SchemaResult {
     this.referencedSchemas = referencedSchemas;
   }
 
-  public static SchemaResult of(Schema schema, Map<String, Schema> referencedSchemas) {
+  @JsonCreator
+  public static SchemaResult of(
+    @JsonProperty("schema") Schema schema,
+    @JsonProperty("referencedSchemas") Map<String, Schema> referencedSchemas
+  ) {
     return new SchemaResult(schema, referencedSchemas);
   }
