+++ openapi-schema-generator/src/main/java/com/github/emm035/openapi/schema/generator/visitors/MapFormatVisitor.java
@@ -8,11 +8,11 @@ import com.fasterxml.jackson.databind.jsonFormatVisitors.JsonMapFormatVisitor;
 import com.github.emm035.openapi.core.v3.references.Referenceable;
 import com.github.emm035.openapi.core.v3.schemas.ObjectSchema;
 import com.github.emm035.openapi.core.v3.schemas.Schema;
-import com.github.emm035.openapi.schema.generator.SchemaGeneratorVisitorWrapper;
-import com.github.emm035.openapi.schema.generator.annotations.Internal;
-import com.github.emm035.openapi.schema.generator.assisted.Extension;
+import com.github.emm035.openapi.schema.generator.annotations.Extension;
+import com.github.emm035.openapi.schema.generator.annotations.internal.Internal;
 import com.github.emm035.openapi.schema.generator.base.Generator;
 import com.github.emm035.openapi.schema.generator.extension.SchemaExtension;
+import com.github.emm035.openapi.schema.generator.visitors.SchemaGeneratorVisitorWrapper;
 import com.google.inject.Inject;
 import com.google.inject.assistedinject.Assisted;
