+++ openapi-schema-generator/src/main/java/com/github/emm035/openapi/schema/generator/visitors/NumberFormatVisitor.java
@@ -9,7 +9,7 @@ import com.github.emm035.openapi.core.v3.schemas.IntegerSchema;
 import com.github.emm035.openapi.core.v3.schemas.NumberSchema;
 import com.github.emm035.openapi.core.v3.schemas.NumericSchema;
 import com.github.emm035.openapi.core.v3.schemas.Schema;
-import com.github.emm035.openapi.schema.generator.assisted.Extension;
+import com.github.emm035.openapi.schema.generator.annotations.Extension;
 import com.github.emm035.openapi.schema.generator.base.Generator;
 import com.github.emm035.openapi.schema.generator.extension.SchemaExtension;
 import com.google.inject.Inject;
