+++ openapi-schema-generator/src/main/java/com/github/emm035/openapi/schema/generator/visitors/SchemaGeneratorVisitorWrapper.java
@@ -1,4 +1,4 @@
-package com.github.emm035.openapi.schema.generator;
+package com.github.emm035.openapi.schema.generator.visitors;
 
 import com.fasterxml.jackson.databind.JavaType;
 import com.fasterxml.jackson.databind.JsonMappingException;
@@ -12,10 +12,11 @@ import com.fasterxml.jackson.databind.jsonFormatVisitors.JsonObjectFormatVisitor
 import com.fasterxml.jackson.databind.jsonFormatVisitors.JsonStringFormatVisitor;
 import com.github.emm035.openapi.core.v3.references.Referenceable;
 import com.github.emm035.openapi.core.v3.schemas.Schema;
-import com.github.emm035.openapi.schema.generator.assisted.SubTypeGenerator;
 import com.github.emm035.openapi.schema.generator.base.Generator;
+import com.github.emm035.openapi.schema.generator.base.Schemas;
 import com.github.emm035.openapi.schema.generator.base.TypeUtils;
 import com.github.emm035.openapi.schema.generator.exceptions.SchemaGeneratorInternalException;
+import com.github.emm035.openapi.schema.generator.nested.SubTypeGenerator;
 import com.github.emm035.openapi.schema.generator.visitors.ArrayFormatVisitor;
 import com.github.emm035.openapi.schema.generator.visitors.BooleanFormatVisitor;
 import com.github.emm035.openapi.schema.generator.visitors.MapFormatVisitor;
@@ -35,6 +36,7 @@ public class SchemaGeneratorVisitorWrapper
   private final ArrayFormatVisitor.Factory arrayFormatVisitorFactory;
   private final MapFormatVisitor.Factory mapFormatVisitorFactory;
   private final SubTypedObjectFormatVisitor.Factory compositeObjectFormatVisitorFactory;
+  private final Schemas schemas;
 
   private Generator schemaGenerator = null;
 
@@ -47,7 +49,8 @@ public class SchemaGeneratorVisitorWrapper
     ArrayFormatVisitor.Factory arrayFormatVisitorFactory,
     MapFormatVisitor.Factory mapFormatVisitorFactory,
     SubTypeGenerator.Factory subTypeGeneratorFactory,
-    SubTypedObjectFormatVisitor.Factory compositeObjectFormatVisitorFactory
+    SubTypedObjectFormatVisitor.Factory compositeObjectFormatVisitorFactory,
+    Schemas schemas
   ) {
     this.stringFormatVisitorFactory = stringFormatVisitorFactory;
     this.numberFormatVisitorFactory = numberFormatVisitorFactory;
@@ -56,6 +59,7 @@ public class SchemaGeneratorVisitorWrapper
     this.arrayFormatVisitorFactory = arrayFormatVisitorFactory;
     this.mapFormatVisitorFactory = mapFormatVisitorFactory;
     this.compositeObjectFormatVisitorFactory = compositeObjectFormatVisitorFactory;
+    this.schemas = schemas;
   }
 
   @Override
