+++ openapi-schema-generator/src/main/java/com/github/emm035/openapi/schema/generator/visitors/SubTypedObjectFormatVisitor.java
@@ -10,15 +10,15 @@ import com.github.emm035.openapi.core.v3.schemas.Discriminator;
 import com.github.emm035.openapi.core.v3.schemas.ObjectSchema;
 import com.github.emm035.openapi.core.v3.schemas.OneOfSchema;
 import com.github.emm035.openapi.core.v3.schemas.Schema;
-import com.github.emm035.openapi.schema.generator.assisted.Extension;
-import com.github.emm035.openapi.schema.generator.assisted.NestedSchemaGenerator;
+import com.github.emm035.openapi.schema.generator.annotations.Extension;
 import com.github.emm035.openapi.schema.generator.assisted.RefFactory;
-import com.github.emm035.openapi.schema.generator.assisted.SubTypeGenerator;
 import com.github.emm035.openapi.schema.generator.base.Generator;
 import com.github.emm035.openapi.schema.generator.base.Schemas;
 import com.github.emm035.openapi.schema.generator.base.TypeUtils;
 import com.github.emm035.openapi.schema.generator.extension.PropertyExtension;
 import com.github.emm035.openapi.schema.generator.extension.SchemaExtension;
+import com.github.emm035.openapi.schema.generator.nested.NestedSchemaGenerator;
+import com.github.emm035.openapi.schema.generator.nested.SubTypeGenerator;
 import com.google.common.base.Predicates;
 import com.google.common.collect.Maps;
 import com.google.inject.Inject;
